+++ doc/RELEASES.md
@@ -2,11 +2,19 @@ Release notes
 =============
 
 
+11.5.2016
+---------
+
+- entity-index matching between device (chassis) and interfaces. You need to match device.physicalIndex and interface[index].physicalIndex
+- device info provides the entity class and more (chassis, psu, module, stack) types of entities. You might want to filter on the desired class, e.g. to get almost the same results as the previous version, extract the entities where physicalClass = "chassis".
+
+
 15.4.2016
 ---------
 
 - add trunk collection with /aj/api/v1/trunk/device.name
 - get the same info attached to each interface with /aj/api/v1/interfaces/device.name?showtrunks=1 
+- configurable port_binding for Flask
 
 
 13.4.2016